refactor(web/user-settings): update the design pattern
