+++ apps/web/src/containers/UserSettings/UserSettings.Profile.tsx
@@ -3,7 +3,7 @@ import { Trans } from 'react-i18next';
 import { createPortal } from 'react-dom';
 
 export default function Profile() {
-  const container = global.document?.getElementById('actions-profile');
+  const container = global.document?.getElementById('profile');
 
   return (
     <>
